pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aisec"
version = "0.1.0"
description = "Deep security analysis framework for autonomous AI agent implementations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Federico Boiero", email = "fboiero@gmail.com" },
]
keywords = [
    "ai-security",
    "llm-security",
    "autonomous-agents",
    "owasp",
    "prompt-injection",
    "docker",
    "security-analysis",
    "privacy",
    "gdpr",
    "habeas-data",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "docker>=7.0.0",
    "jinja2>=3.1.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
pdf = ["weasyprint>=62.0"]
network = ["scapy>=2.5.0"]
secrets = ["detect-secrets>=1.4.0"]
all = ["aisec[pdf,network,secrets]"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
aisec = "aisec.cli.app:app"

[project.entry-points."aisec.plugins"]

[project.urls]
Homepage = "https://github.com/fboiero/AiSec"
Documentation = "https://github.com/fboiero/AiSec/tree/main/docs"
Repository = "https://github.com/fboiero/AiSec"
Issues = "https://github.com/fboiero/AiSec/issues"
Changelog = "https://github.com/fboiero/AiSec/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/aisec"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
